Tweak .stream() API
