+++ tests/client/test_async_client.py
@@ -61,7 +61,7 @@ async def test_post_json(server):
 @pytest.mark.usefixtures("async_environment")
 async def test_stream_response(server):
     async with httpx.AsyncClient() as client:
-        async with client.stream("GET", server.url) as response:
+        async with await client.stream("GET", server.url) as response:
             body = await response.aread()
 
     assert response.status_code == 200
@@ -72,7 +72,7 @@ async def test_stream_response(server):
 @pytest.mark.usefixtures("async_environment")
 async def test_access_content_stream_response(server):
     async with httpx.AsyncClient() as client:
-        async with client.stream("GET", server.url) as response:
+        async with await client.stream("GET", server.url) as response:
             pass
 
     assert response.status_code == 200
